Move all cache related code to artifact_cache
